Try to make collision better
Have I ever said that collision is hard? It doesn't help when I'm tired and
trying to get the implementation done as quickly as possible. What I should do
is code a 2d platformer prototype. That will give me the basics. Then I can
move it to 3d, and then move that here.

Collision is hard.